Matin/Update: Dashboard colors
