+++ packages/components/src/components/button-toggle/button-highlight.jsx
@@ -1,10 +1,9 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-const Highlight = ({ has_rounded_button, highlight_color, left, width }) => {
+const Highlight = ({ has_rounded_button, left, width }) => {
     const border_radius_size = '4px';
     const highlight_style = {
-        backgroundColor: `var(--button-toggle-${highlight_color || 'secondary'})`,
         left: 0,
         transform: `translate3d(${left}px, 0, 0)`,
         width,
