+++ packages/components/src/components/button-toggle/button-toggle.jsx
@@ -5,17 +5,7 @@ import HighlightWrapper from './button-highlight-wrapper.jsx';
 import Button from '../button/button.jsx';
 import Counter from '../counter/counter.jsx';
 
-const ButtonToggle = ({
-    buttons_arr,
-    className,
-    has_rounded_button,
-    highlight_color,
-    id,
-    is_animated,
-    name,
-    onChange,
-    value,
-}) => {
+const ButtonToggle = ({ buttons_arr, className, has_rounded_button, id, is_animated, name, onChange, value }) => {
     const changeValue = selected_value => {
         if (value === selected_value) return;
         onChange({ target: { value: selected_value, name } });
@@ -47,9 +37,7 @@ const ButtonToggle = ({
     return (
         <div id={id} className={classNames('dc-button-menu', className)}>
             {is_animated ? (
-                <HighlightWrapper has_rounded_button={has_rounded_button} highlight_color={highlight_color}>
-                    {menu}
-                </HighlightWrapper>
+                <HighlightWrapper has_rounded_button={has_rounded_button}>{menu}</HighlightWrapper>
             ) : (
                 <React.Fragment>{menu}</React.Fragment>
             )}
