+++ packages/dashboard/src/components/pages/home/home.scss
@@ -22,10 +22,12 @@
         }
 
         &--highlight {
-            color: #{$color-white};
-            background-color: #{$color-apple-blue};
-            border: none;
-            height: 30px;
+            height: 3rem;
+            transition: unset;
+            transform: translate3d(0, 1.6rem, 0);
+            border-left: unset;
+            border-right: 4px solid var(--brand-red-coral);
+            border-radius: 0 !important;
         }
 
         &__link {
@@ -69,7 +71,6 @@
     &__header,
     &__header-group {
         &--active &__link {
-            color: var(--text-colored-background);
             font-size: 1.5rem;
             font-weight: bold;
         }
