--- packages/dashboard/src/components/template/accounts-button-toggle/accounts-button-toggle.scss
@@ -1,24 +1,0 @@
-.button-toggle {
-    &__container {
-        padding: 0.8rem;
-        background-color: var(--general-section-1);
-        border-radius: 0.8rem;
-    }
-    &__items {
-        .dc-button-menu {
-            &__wrapper {
-                background-color: unset !important;
-            }
-            &__button {
-                width: 9.6rem;
-                height: 3.2rem;
-                span {
-                    font-size: var(--text-size-s);
-                }
-                &--active span {
-                    color: var(--text-colored-background);
-                }
-            }
-        }
-    }
-}
