--- packages/dashboard/src/components/template/accounts-button-toggle/index.ts
@@ -1,4 +1,0 @@
-import AccountsToggleButton from './accounts-button-toggle';
-import './accounts-button-toggle.scss';
-
-export default AccountsToggleButton;
