+++ packages/dashboard/src/components/template/accounts-toggle-button/accounts-toggle-button.scss
@@ -1,0 +1,26 @@
+.button-toggle {
+    &__container {
+        display: inline-block;
+        width: fit-content;
+        padding: 0.8rem;
+        background-color: var(--general-section-1);
+        border-radius: 0.8rem;
+    }
+    &__items {
+        .dc-button-menu {
+            &__wrapper {
+                background-color: unset !important;
+            }
+            &__button {
+                width: 9.6rem;
+                height: 3.2rem;
+                span {
+                    font-size: var(--text-size-s);
+                }
+            }
+            &--highlight {
+                background-color: var(--state-normal);
+            }
+        }
+    }
+}
