+++ packages/dashboard/src/components/template/accounts-toggle-button/index.ts
@@ -1,0 +1,4 @@
+import AccountsToggleButton from './accounts-toggle-button';
+import './accounts-toggle-button.scss';
+
+export default AccountsToggleButton;
