+++ packages/dashboard/src/components/template/app-to-trade/app-to-trade.scss
@@ -39,14 +39,14 @@
             margin-top: 1.6rem;
         }
         &-link {
-            color: $color-apple-blue;
+            color: $color-blue-4;
             text-decoration: none;
             font-weight: bold;
             font-size: 12px;
             margin: auto 2.4rem auto 0;
         }
         .dc-btn--primary {
-            background-color: $color-apple-blue;
+            background-color: $color-blue-4;
         }
     }
     &__benefits-wrapper {
