+++ packages/shared/src/styles/constants.scss
@@ -42,7 +42,6 @@ $color-green-3: #00a79e;
 $color-green-4: #008079;
 $color-orange: #ff6444;
 $color-yellow: #ffad3a;
-$color-apple-blue: #0677af;
 $color-blue: #377cfc;
 $color-blue-1: #0dc2e7;
 $color-blue-2: #2a3052;
