+++ packages/shared/src/styles/themes.scss
@@ -130,9 +130,6 @@
         --status-info: #{$color-blue};
         --status-colored-background: #{$color-white};
         // Dashboard
-        --dashboard-primary: #{$color-blue-3};
-        --dashboard-hover: #{$color-blue-4};
-        --dashboard-alpha: #{$alpha-color-blue-2};
         --dashboard-swap-free: #{$gradient-color-black-3};
         --dashboard-app: #{$gradient-color-black-4};
         // Payment methods
@@ -159,7 +156,7 @@
         --gradient-right-edge: #{$gradient-color-white};
         // Badge
         --badge-white: #{$color-white};
-        --badge-blue: #{$color-apple-blue};
+        --badge-blue: #{$color-blue-4};
         --badge-violet: #{$color-blue-2};
         --badge-green: #{$color-green-3};
     }
@@ -230,10 +227,6 @@
         --status-success: #{$color-green-3};
         --status-info: #{$color-blue};
         --status-colored-background: #{$color-white};
-        // Dashboard
-        --dashboard-primary: #{$color-blue-2};
-        --dashboard-hover: #{$color-blue-3};
-        --dashboard-alpha: #{$alpha-color-blue-2};
         // Transparentize
         --transparent-success: #{$alpha-color-green-2};
         --transparent-info: #{$alpha-color-blue-1};
@@ -244,7 +237,7 @@
         --gradient-right-edge: #{$gradient-color-black};
         // Badge
         --badge-white: #{$color-white};
-        --badge-blue: #{$color-apple-blue};
+        --badge-blue: #{$color-blue-4};
         --badge-violet: #{$color-blue-2};
         --badge-green: #{$color-green-3};
     }
